+++ packages/wonder-blocks-dropdown/src/util/constants.js
@@ -19,7 +19,6 @@ export const selectDropdownStyle = {
 // Note that these can be overridden by the provided style if needed.
 export const filterableDropdownStyle = {
     minHeight: 100,
-    maxHeight: 384,
 };
 
 export const searchInputStyle = {