+++ packages/wonder-blocks-search-field/src/components/search-field.js
@@ -1,10 +1,9 @@
 // @flow
 import * as React from "react";
-import * as ReactDOM from "react-dom";
 import {StyleSheet} from "aphrodite";
 
 import {styles as typographyStyles} from "@khanacademy/wonder-blocks-typography";
-import {View} from "@khanacademy/wonder-blocks-core";
+import {View, IDProvider} from "@khanacademy/wonder-blocks-core";
 import IconButton from "@khanacademy/wonder-blocks-icon-button";
 import {TextField} from "@khanacademy/wonder-blocks-form";
 import Icon, {icons} from "@khanacademy/wonder-blocks-icon";
@@ -23,9 +22,10 @@ type Props = {|
     clearAriaLabel?: string,
 
     /**
-     * The unique identifier for the input.
+     * The unique identifier for the input. If one is not provided,
+     * a unique id will be generated.
      */
-    id: string,
+    id?: string,
 
     /**
      * The text input value.
@@ -107,78 +107,100 @@ type Props = {|
  * />
  * ```
  */
-export default function SearchField(props: Props): React.Node {
-    const {
-        clearAriaLabel = defaultLabels.clearSearch,
-        disabled = false,
-        light = false,
-        id,
-        value,
-        placeholder,
-        style,
-        testId,
-        onClick,
-        onChange,
-        onFocus,
-        onBlur,
-        ...otherProps
-    } = props;
+const SearchField: React.AbstractComponent<Props, HTMLInputElement> =
+    React.forwardRef<Props, HTMLInputElement>((props: Props, ref) => {
+        const {
+            clearAriaLabel = defaultLabels.clearSearch,
+            disabled = false,
+            light = false,
+            id,
+            value,
+            placeholder,
+            style,
+            testId,
+            onClick,
+            onChange,
+            onFocus,
+            onBlur,
+            ...otherProps
+        } = props;
 
-    const handleClear: () => void = () => {
-        // Empty the search text.
-        onChange("");
+        // We can't just use ref.current to clear the input because ref isn't
+        // always being passed in, so we use an innerRef to allow the
+        // handleClear() function to focus on the input element ref.
+        const innerRef = React.useRef<?HTMLInputElement>(null);
 
-        // Focus back on the text field since the clear button disappears after
-        // the field is cleared.
-        const currentField = (ReactDOM.findDOMNode(
-            document.getElementById(id),
-        ): any);
-        currentField.focus();
-    };
+        const handleClear: () => void = () => {
+            // Empty the search text.
+            onChange("");
 
-    const maybeRenderClearIconButton: () => React.Node = () => {
-        if (!value.length) {
-            return null;
-        }
+            // Focus back on the text field since the clear button disappears after
+            // the field is cleared.
+            innerRef?.current?.focus();
+        };
+
+        const maybeRenderClearIconButton: () => React.Node = () => {
+            if (!value.length) {
+                return null;
+            }
+
+            return (
+                <IconButton
+                    icon={icons.dismiss}
+                    kind="tertiary"
+                    onClick={handleClear}
+                    style={styles.dismissIcon}
+                    aria-label={clearAriaLabel}
+                />
+            );
+        };
 
         return (
-            <IconButton
-                icon={icons.dismiss}
-                kind="tertiary"
-                onClick={handleClear}
-                style={styles.dismissIcon}
-                aria-label={clearAriaLabel}
-            />
+            <IDProvider id={id} scope="search-field">
+                {(uniqueId) => (
+                    <View
+                        onClick={onClick}
+                        style={[styles.inputContainer, style]}
+                    >
+                        <Icon
+                            icon={icons.search}
+                            size="medium"
+                            color={Color.offBlack64}
+                            style={styles.searchIcon}
+                            aria-hidden="true"
+                        />
+                        <TextField
+                            id={`${uniqueId}-field`}
+                            type="text"
+                            disabled={disabled}
+                            light={light}
+                            onChange={onChange}
+                            onFocus={onFocus}
+                            onBlur={onBlur}
+                            placeholder={placeholder}
+                            ref={(node) => {
+                                // We have to set the value of both refs to
+                                // the HTMLInputElement from TextField.
+                                if (ref) {
+                                    // $FlowIgnore[prop-missing]
+                                    ref.current = node;
+                                }
+                                innerRef.current = node;
+                            }}
+                            value={value}
+                            style={[
+                                styles.inputStyleReset,
+                                typographyStyles.LabelMedium,
+                            ]}
+                            testId={testId}
+                            {...otherProps}
+                        />
+                        {maybeRenderClearIconButton()}
+                    </View>
+                )}
+            </IDProvider>
         );
-    };
-
-    return (
-        <View onClick={onClick} style={[styles.inputContainer, style]}>
-            <Icon
-                icon={icons.search}
-                size="medium"
-                color={Color.offBlack64}
-                style={styles.searchIcon}
-                aria-hidden="true"
-            />
-            <TextField
-                id={id}
-                type="text"
-                disabled={disabled}
-                light={light}
-                onChange={onChange}
-                onFocus={onFocus}
-                onBlur={onBlur}
-                placeholder={placeholder}
-                value={value}
-                style={[styles.inputStyleReset, typographyStyles.LabelMedium]}
-                testId={testId}
-                {...otherProps}
-            />
-            {maybeRenderClearIconButton()}
-        </View>
-    );
-}
+    });
 
 const styles = StyleSheet.create({
     inputContainer: {
@@ -213,3 +235,5 @@ const styles = StyleSheet.create({
         paddingRight: Spacing.large_24 + Spacing.medium_16,
     },
 });
+
+export default SearchField;