+++ packages/wonder-blocks-breadcrumbs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/wonder-blocks-breadcrumbs
 
+## 1.0.30
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-core@4.3.1
+-   @khanacademy/wonder-blocks-typography@1.1.31
+
 ## 1.0.29
 
 ### Patch Changes