+++ packages/wonder-blocks-cell/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @khanacademy/wonder-blocks-cell
 
+## 2.2.3
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-clickable@2.2.6
+-   @khanacademy/wonder-blocks-core@4.3.1
+-   @khanacademy/wonder-blocks-layout@1.4.9
+-   @khanacademy/wonder-blocks-typography@1.1.31
+
+## 2.2.2
+
+### Patch Changes
+
+-   f74b7379: Fixes `CellCore` to set the styles properly (removed `marginLeft: auto` from accessory and expanded the content wrapper).
+
 ## 2.2.1
 
 ### Patch Changes