+++ packages/wonder-blocks-search-field/CHANGELOG.md
@@ -1,10 +1,32 @@
 # @khanacademy/wonder-blocks-search-field
 
+## 1.0.3
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-core@4.3.1
+-   @khanacademy/wonder-blocks-icon@1.2.27
+-   @khanacademy/wonder-blocks-icon-button@3.4.6
+-   @khanacademy/wonder-blocks-typography@1.1.31
+
+## 1.0.2
+
+### Patch Changes
+
+-   9ecc853d: Wrap the field in a UniqueIDProvider
+
+## 1.0.1
+
+### Patch Changes
+
+-   9f8ca907: Added forwardRef so it can be used by dropdown
+
 ## 1.0.0
+
 ### Major Changes
 
-- 2d9198ff: Create SearchField component
+-   2d9198ff: Create SearchField component
 
 ### Minor Changes
 
-- 80aebdc2: Adds `wonder-blocks-search-field` package
+-   80aebdc2: Adds `wonder-blocks-search-field` package